requirements.txt
pandas
mysql-connector-python
flask
blinker==1.6.2
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
numpy==1.25.2
pandas==2.0.3
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.6
